pyproject.toml
[tool.poetry]
name = "pytrengo"
version = "0.1.0"
description = ""
authors = ["Baptiste Fontaine <baptiste@bixoto.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
api-session = "^1.4.1"
python-dotenv = "^1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
